pyproject.toml
[tool.poetry]
name = "poetry add fastapi"
version = "0.1.0"
description = ""
authors = ["Darkstrouk <Avborodin@edu.hse.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
httpx = "^0.27.0"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

